requirements.txt
streamlit
gspread
pandas
plotly
numpy
google-auth
